PENDING: refactor(ti): Restructure firewall configuration code on AM62L
Commit 11e6effa9 ("PENDING: feat(ti): Simplify firewall configuration
on AM62L") disables the offending background firewall configurations
that are enabled by ROM for boot phase, and are not needed after boot
is done.
Similar configuration needs to be done when SoC resumes from RTC + DDR
low power mode.
Move firewall related functions to a board-specific location so that these
configurations can be done during resume as well.
